Refactor/session cookies

+++ internal/data/session.go
@@ -8,10 +8,29 @@ import (
 
 const (
 	createSessionQuery = `
-  INSERT INTO sessions (user_id, token, user_agent, ip_address, created_at, expires_at) 
-  VALUES ($1, $2, $3, $4, $5, $6)`
-	selectSessionByTokenQuery = "SELECT * FROM sessions WHERE token = $1"
-	deleteSessionByTokenQuery = "DELETE FROM sessions WHERE token = $1"
+  INSERT INTO sessions (user_id, user_agent, ip_address, created_at, expires_at) 
+  VALUES ($1, $2, $3, $4, $5)
+  RETURNING *`
+	// TODO: this is where we are at. Fetch the session with the user with a join query
+	selectSessionWithUserQuery = `
+  SELECT 
+    sessions.id,
+    sessions.user_id,
+    sessions.user_agent,
+    sessions.ip_address,
+    sessions.created_at,
+    sessions.expires_at,
+    users.id,
+    users.firstname,
+    users.lastname,
+    users.email,
+    users.deleted_at,
+    users.verified
+  FROM sessions 
+  JOIN users ON sessions.user_id = users.id
+  WHERE sessions.id = $1 AND sessions.user_id = $2
+  `
+	deleteSessionByTokenQuery = "DELETE FROM sessions WHERE token = $1 AND user_id = $2"
 )
 
 type SessionRepository struct {
@@ -22,28 +41,55 @@ func NewSessionRepository(db *DBPool) *SessionRepository {
 	return &SessionRepository{db: db}
 }
 
-func (r *SessionRepository) Insert(ctx context.Context, s *models.Session) error {
-	_, err := r.db.Query(ctx, createSessionQuery, s.UserID, s.Token, s.UserAgent, s.IpAddr, s.CreatedAt, s.ExpiresAt)
-	return err
+func (r *SessionRepository) Insert(ctx context.Context, s *models.Session) (*models.Session, error) {
+	var session models.Session
+	err := r.db.QueryRow(
+		ctx,
+		createSessionQuery,
+		s.UserID,
+		s.UserAgent,
+		s.IpAddr,
+		s.CreatedAt,
+		s.ExpiresAt).
+		Scan(
+			&session.ID,
+			&session.UserID,
+			&session.UserAgent,
+			&session.IpAddr,
+			&session.CreatedAt,
+			&session.ExpiresAt,
+		)
+	return &session, err
 }
 
-func (r *SessionRepository) GetByToken(ctx context.Context, token models.SessionToken) (*models.Session, error) {
-	row := r.db.QueryRow(ctx, selectSessionByTokenQuery, token)
+func (r *SessionRepository) GetWithUser(ctx context.Context, sessionID models.ID, userID models.ID) (*models.Session, *models.User, error) {
+	row := r.db.QueryRow(ctx, selectSessionWithUserQuery, sessionID, userID)
 
 	var s models.Session
-	err := row.Scan(&s.ID, &s.UserID, &s.Token, &s.UserAgent, &s.IpAddr, &s.CreatedAt, &s.ExpiresAt)
+	var u models.User
+	err := row.Scan(
+		&s.ID,
+		&s.UserID,
+		&s.UserAgent,
+		&s.IpAddr,
+		&s.CreatedAt,
+		&s.ExpiresAt,
+
+		&u.ID,
+		&u.Firstname,
+		&u.Lastname,
+		&u.Email,
+		&u.DeletedAt,
+		&u.Verified,
+	)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
-	return &s, nil
+	return &s, &u, nil
 }
 
-func (r *SessionRepository) Update(ctx context.Context, token models.SessionToken, session *models.Session) error {
-	panic("session repo update not implemented")
-}
-
-func (r *SessionRepository) Delete(ctx context.Context, token models.SessionToken) error {
-	_, err := r.db.Exec(ctx, deleteSessionByTokenQuery, token)
+func (r *SessionRepository) Delete(ctx context.Context, sessionID models.ID, userID models.ID) error {
+	_, err := r.db.Exec(ctx, deleteSessionByTokenQuery, sessionID, userID)
 	return err
 }
